Split default title render test into separate file
meta:sha:1a3aea097475c2ff17848b1309b82f8e0e77b29d
